% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms.R
\name{doubledeepms}
\alias{doubledeepms}
\title{doubledeepms}
\usage{
doubledeepms(startStage = 1, stopStage = 0,
  base_dir = "/users/project/prj004631/afaure/DMS/Results/doubledeepms_proj",
  rerun_raw = F, Ncores = 1)
}
\arguments{
\item{startStage}{Start at a specified analysis stage (default:1)}

\item{stopStage}{Stop at a specified analysis stage (default:0 i.e. no stop condition)}

\item{base_dir}{Base directory for all output file (default:NB private CRG server path; change accordingly)}

\item{rerun_raw}{re-run raw data preprocessing (default:F)}

\item{Ncores}{# cores available to parallelize leave-one-out cross validation workflow (stage 4)}
}
\value{
Nothing
}
\description{
Main analysis script.
}
