% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_3did_comparisons.R
\name{doubledeepms_3did_comparisons}
\alias{doubledeepms_3did_comparisons}
\title{doubledeepms_3did_comparisons}
\usage{
doubledeepms_3did_comparisons(input_file, threedid_file,
  alignment_file_list, threedid_domain_name_list, position_offset_list,
  outpath, colour_scheme, execute = TRUE)
}
\arguments{
\item{input_file}{path to input file (required)}

\item{threedid_file}{3did flat file (required)}

\item{alignment_file_list}{alignment file list (required)}

\item{threedid_domain_name_list}{3did domain name list (required)}

\item{position_offset_list}{position offset list (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot free energy heatmaps.
}
