% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__aa_properties_pca.R
\name{doubledeepms__aa_properties_pca}
\alias{doubledeepms__aa_properties_pca}
\title{doubledeepms__aa_properties_pca}
\usage{
doubledeepms__aa_properties_pca(aa_properties_file,
  selected_identifiers = NULL, return_evidences = F,
  return_matrix_only = F)
}
\arguments{
\item{aa_properties_file}{path to amino acid properties file (required)}

\item{selected_identifiers}{path to file with selected subset of identifiers}

\item{return_evidences}{return AA properties evidences (default:F)}

\item{return_matrix_only}{only return AA properties matrix (default:F)}
}
\value{
results from PCA
}
\description{
Create results folder for analysis script plots and saved objects.
}
