% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__aa_properties_pca_WT_loadings.R
\name{doubledeepms__aa_properties_pca_WT_loadings}
\alias{doubledeepms__aa_properties_pca_WT_loadings}
\title{doubledeepms__aa_properties_pca_WT_loadings}
\usage{
doubledeepms__aa_properties_pca_WT_loadings(input_dt, aa_properties_file,
  selected_identifiers = NULL)
}
\arguments{
\item{input_dt}{data.table with double mutant codes (required)}

\item{aa_properties_file}{path to amino acid properties file (required)}

\item{selected_identifiers}{path to file with selected subset of identifiers}
}
\value{
results from PCA
}
\description{
Create results folder for analysis script plots and saved objects.
}
