% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/doubledeepms__allosteric_mutations_scatterplot.R
\name{doubledeepms__allosteric_mutations_scatterplot}
\alias{doubledeepms__allosteric_mutations_scatterplot}
\title{doubledeepms__allosteric_mutations_scatterplot}
\usage{
doubledeepms__allosteric_mutations_scatterplot(input_dt, outpath,
  colour_scheme)
}
\arguments{
\item{input_dt}{input data table (required)}

\item{outpath}{plot output path (required)}

\item{colour_scheme}{colour scheme file (required)}
}
\value{
Allosteric mutations
}
\description{
Allosteric mutations scatterplot.
}
