% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__calculate_angle.R
\name{doubledeepms__calculate_angle}
\alias{doubledeepms__calculate_angle}
\title{doubledeepms__calculate_angle}
\usage{
doubledeepms__calculate_angle(a, b)
}
\arguments{
\item{a}{vector a (required)}

\item{b}{vector b (required)}
}
\value{
angle in degrees
}
\description{
Calculate angle between vectors in degrees
}
