% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__create_dir.R
\name{doubledeepms__create_dir}
\alias{doubledeepms__create_dir}
\title{doubledeepms__create_dir}
\usage{
doubledeepms__create_dir(doubledeepms_dir, execute = TRUE,
  message = NULL, overwrite_dir = TRUE)
}
\arguments{
\item{doubledeepms_dir}{directory path string (required)}

\item{execute}{whether or not the system command will be executed (required)}

\item{message}{message string (optional, default: NULL i.e. no message displayed)}

\item{overwrite_dir}{delete directory if already exists (optional, default: TRUE)}
}
\value{
Nothing
}
\description{
Create results folder for analysis script plots and saved objects.
}
