% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__define_confident_free_energies.R
\name{doubledeepms__define_confident_free_energies}
\alias{doubledeepms__define_confident_free_energies}
\title{doubledeepms__define_confident_free_energies}
\usage{
doubledeepms__define_confident_free_energies(input_dt, report_outpath,
  folding_energy_max_sd = 1/(1.96 * 2), binding_energy_max_sd = 1/(1.96
  * 2))
}
\arguments{
\item{input_dt}{data.table with model free energy estimates (required)}

\item{report_outpath}{output path for scatterplots (required)}

\item{folding_energy_max_sd}{maximum folding energy standard deviation (default:1/(1.96*2))}

\item{binding_energy_max_sd}{maximum binding energy standard deviation (default:1/(1.96*2))}
}
\value{
data.table with columns indicating confident free energies
}
\description{
Call confident ddGs.
}
