% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__format_dir.R
\name{doubledeepms__format_dir}
\alias{doubledeepms__format_dir}
\title{doubledeepms__format_dir}
\usage{
doubledeepms__format_dir(dir_suffix, stagenum, base_dir, no_digits = 3)
}
\arguments{
\item{dir_suffix}{directory suffix (required)}

\item{stagenum}{stage number (required)}

\item{base_dir}{base directory (required)}

\item{no_digits}{number of digits to use for stage number}
}
\value{
Absolute directory character string
}
\description{
Plot fitness versus growth rate.
}
