% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__fraction_folded.R
\name{doubledeepms__fraction_folded}
\alias{doubledeepms__fraction_folded}
\title{doubledeepms__fraction_folded}
\usage{
doubledeepms__fraction_folded(folding_energy, RT = 0.001987 * (273 + 24))
}
\arguments{
\item{folding_energy}{folding energies for 2-state model (required)}

\item{RT}{constant (default:0.001987*(273+24))}
}
\value{
Fraction folded
}
\description{
Calculate fraction folded in 2-state model.
}
