% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__get_model_results.R
\name{doubledeepms__get_model_results}
\alias{doubledeepms__get_model_results}
\title{doubledeepms__get_model_results}
\usage{
doubledeepms__get_model_results(input_folder, input_dt,
  n_bootstraps = 10, bootstrap_prefix = "boot_", RT = 0.001987 * (273
  + 24))
}
\arguments{
\item{input_folder}{path to input folder with model results (required)}

\item{input_dt}{data.table with model free energy estimates (required)}

\item{n_bootstraps}{number of bootstraps (default:10)}

\item{bootstrap_prefix}{bootstrap subfolder prefix (default:"boot_")}

\item{RT}{constant (default:0.001987*(273+24))}
}
\value{
data.table with model results
}
\description{
Get model results and coefficients.
}
