% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__mann_whitney_U_wrapper.R
\name{doubledeepms__mann_whitney_U_wrapper}
\alias{doubledeepms__mann_whitney_U_wrapper}
\title{doubledeepms__mann_whitney_U_wrapper}
\usage{
doubledeepms__mann_whitney_U_wrapper(vals1, vals2)
}
\arguments{
\item{vals1}{sample 1 (required)}

\item{vals2}{sample 2 (required)}
}
\value{
named vector with effect size (AUC) and p-value obtained from coin package
}
\description{
Mann-Whitney U test
}
