% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__mann_whitney_U_wrapper_rand.R
\name{doubledeepms__mann_whitney_U_wrapper_rand}
\alias{doubledeepms__mann_whitney_U_wrapper_rand}
\title{doubledeepms__mann_whitney_U_wrapper_rand}
\usage{
doubledeepms__mann_whitney_U_wrapper_rand(vals1, vals2, sd1, sd2,
  n = 100, strictly_positive = TRUE)
}
\arguments{
\item{vals1}{sample 1 (required)}

\item{vals2}{sample 2 (required)}

\item{sd1}{sample 1 sd (required)}

\item{sd2}{sample 2 sd (required)}

\item{n}{number of randomisations (default:100)}

\item{strictly_positive}{values are strictly positive (default:TRUE)}
}
\value{
named vector with effect size (AUC) and p-value obtained from coin package
}
\description{
Mann-Whitney U test
}
