% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/doubledeepms__minimum_interchain_distances_from_PDB.R
\name{doubledeepms__minimum_interchain_distances_from_PDB}
\alias{doubledeepms__minimum_interchain_distances_from_PDB}
\title{doubledeepms__minimum_interchain_distances_from_PDB}
\usage{
doubledeepms__minimum_interchain_distances_from_PDB(input_file,
  chain_query = "A", chain_target = "B")
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{chain_query}{query chain id (default:A)}

\item{chain_target}{target chain id (default:B)}
}
\value{
data.table with minimum inter-chain (side-chain) heavy atom distances
}
\description{
Calculate minimum inter-chain (side-chain) heavy atom distances.
}
