% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__mochi__construct_feature_matrix.R
\name{doubledeepms__mochi__construct_feature_matrix}
\alias{doubledeepms__mochi__construct_feature_matrix}
\title{doubledeepms__mochi__construct_feature_matrix}
\usage{
doubledeepms__mochi__construct_feature_matrix(input_dt, prefix = "",
  mut_ids = NULL)
}
\arguments{
\item{input_dt}{Input data.table (required)}

\item{prefix}{Prefix for feature names (default:empty string)}

\item{mut_ids}{Mutation id subset (default:NULL)}
}
\value{
Feature matrix data.table
}
\description{
Contruct feature matrix for model fitting.
}
