% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__mochi__encode_genotypes.R
\name{doubledeepms__mochi__encode_genotypes}
\alias{doubledeepms__mochi__encode_genotypes}
\title{doubledeepms__mochi__encode_genotypes}
\usage{
doubledeepms__mochi__encode_genotypes(input_dt,
  sequenceType = "nucleotide", remove_constant = T)
}
\arguments{
\item{input_dt}{Input data.table (required)}

\item{sequenceType}{Sequence type: 'nucleotide' or 'aminoacid' (default:nucleotide)}

\item{remove_constant}{Remove invariable positions from encoded genotype (default:T)}
}
\value{
Encoded genotypes dictionary
}
\description{
Encode genotypes mapping id (comma-separated mutation string) to var (single character variant string where WT=0) for variable positions only.
}
