% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__mochi__fit_tmodel_3state.R
\name{doubledeepms__mochi__fit_tmodel_3state}
\alias{doubledeepms__mochi__fit_tmodel_3state}
\title{doubledeepms__mochi__fit_tmodel_3state}
\usage{
doubledeepms__mochi__fit_tmodel_3state(fitness_folding_inpath = NULL,
  fitness_binding_inpath = NULL, id_subset_folding_inpath = "all",
  id_subset_binding_inpath = "all", mut_order_subset_train = "1,2",
  mut_order_subset_valid = "2", mut_order_subset_obs = "union",
  min_multiplicity_valid = 2, fit_type = "fb",
  sequence_type = "aminoacid", subsample_prop = 1, mochi_script,
  mochi_outpath, num_epochs = 1000, num_epochs_grid = 100,
  validation_set_proportion = 0.1, num_resamplings = 10,
  num_samples = "128,256,512,1024",
  learning_rate = "0.0001,0.001,0.01,0.1", job_number = 1)
}
\arguments{
\item{fitness_folding_inpath}{Folding fitness DiMSum RData file (default:NULL)}

\item{fitness_binding_inpath}{Folding fitness DiMSum RData file (default:NULL)}

\item{id_subset_folding_inpath}{Folding coefficient id subset file (default:all)}

\item{id_subset_binding_inpath}{Folding coefficient id subset file (default:all)}

\item{mut_order_subset_train}{Mutation order subset for training data, in addition to WT (default:1,2)}

\item{mut_order_subset_valid}{Mutation order subset for validation data, in addition to WT (default:2)}

\item{mut_order_subset_obs}{Mutation order subset for all observed data, in addition to WT (default:union of mutOrderSubsetTrain and mutOrderSubsetValid)}

\item{min_multiplicity_valid}{Minimum multiplicity for validation data (default:2)}

\item{fit_type}{Fit type, either fb=folding+binding, f=folding, b=binding (default:fb)}

\item{sequence_type}{Sequence type (default:aminoacid)}

\item{subsample_prop}{Doubles subsample proportion (default:1)}

\item{mochi_script}{MoCHI script (required)}

\item{mochi_outpath}{Output folder (required)}

\item{num_epochs}{Number of epochs (default:1000)}

\item{num_epochs_grid}{Number of epochs for gridsearch (default:100)}

\item{validation_set_proportion}{Validation set proportion (default:0.1)}

\item{num_resamplings}{Number of random resamplings of target values (default:10)}

\item{num_samples}{Number of samples per gradient update (default:128,256,512,1024)}

\item{learning_rate}{Learning rate (default:0.0001,0.001,0.01,0.1)}

\item{job_number}{Job number (default:1)}
}
\value{
Nothing
}
\description{
Fit 3-state thermodynamic model to ddPCA data.
}
