% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/doubledeepms__mochi__remove_constant_positions.R
\name{doubledeepms__mochi__remove_constant_positions}
\alias{doubledeepms__mochi__remove_constant_positions}
\title{doubledeepms__mochi__remove_constant_positions}
\usage{
doubledeepms__mochi__remove_constant_positions(input_strings)
}
\arguments{
\item{input_strings}{vector of sequence strings (required)}
}
\value{
Vector of sequence strings with invariable positions removed
}
\description{
Remove invariable positions from sequence strings.
}
