% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__persite_energy_vs_distance_plot.R
\name{doubledeepms__persite_energy_vs_distance_plot}
\alias{doubledeepms__persite_energy_vs_distance_plot}
\title{doubledeepms__persite_energy_vs_distance_plot}
\usage{
doubledeepms__persite_energy_vs_distance_plot(input_dt, outpath,
  colour_scheme, trait_name = "folding")
}
\arguments{
\item{input_dt}{processed data table with DiMSum fitness scores (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{trait_name}{trait name (default:folding)}
}
\value{
Nothing
}
\description{
Create results folder for analysis script plots and saved objects.
}
