% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_additive_trait_folding.R
\name{doubledeepms__plot_additive_trait_folding}
\alias{doubledeepms__plot_additive_trait_folding}
\title{doubledeepms__plot_additive_trait_folding}
\usage{
doubledeepms__plot_additive_trait_folding(input_dt, report_outpath)
}
\arguments{
\item{input_dt}{data.table with model free energy estimates (required)}

\item{report_outpath}{output path for scatterplots (required)}
}
\value{
Nothing
}
\description{
Plot folding additive trait.
}
