% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_fitness_replicates_cor.R
\name{doubledeepms__plot_fitness_replicates_cor}
\alias{doubledeepms__plot_fitness_replicates_cor}
\title{doubledeepms__plot_fitness_replicates_cor}
\usage{
doubledeepms__plot_fitness_replicates_cor(input_dt, outpath, colour_scheme)
}
\arguments{
\item{input_dt}{list of folder paths to fitness data (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}
}
\value{
Nothing
}
\description{
Plot fitness replicate correlations.
}
