% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_growth_validations.R
\name{doubledeepms__plot_growth_validations}
\alias{doubledeepms__plot_growth_validations}
\title{doubledeepms__plot_fitness_replicates_cor}
\usage{
doubledeepms__plot_growth_validations(fitness_path, outpath, val_inpath,
  execute = TRUE, message = NULL)
}
\arguments{
\item{fitness_path}{list of folder paths to fitness data (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{val_inpath}{path to experimental fitness validations (required)}

\item{execute}{whether or not the system command will be executed (required)}
}
\value{
Nothing
}
\description{
Create results folder for analysis script plots and saved objects.
}
