% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_heatmap.R
\name{doubledeepms__plot_heatmap}
\alias{doubledeepms__plot_heatmap}
\title{doubledeepms__plot_heatmap}
\usage{
doubledeepms__plot_heatmap(input_dt, variable_name, metric_name,
  output_file, width = 10, height = 4, plot_title = "",
  colour_clip = 4, colour_low = "blue", colour_high = "red",
  colour_mid = "lightgrey", colour_mid_text = "grey")
}
\arguments{
\item{input_dt}{input data.table (required)}

\item{variable_name}{variable name for heatmap cells (required)}

\item{metric_name}{input data.table (required)}

\item{output_file}{plot output file path (required)}

\item{width}{plot width in inches (default:10)}

\item{height}{plot height in inches (default:4)}

\item{plot_title}{main title for plot (default:empty)}

\item{colour_clip}{maximum absolute value of colour scale (default:4)}

\item{colour_low}{colour scale lower limit colour (default:blue)}

\item{colour_high}{colour scale upper limit colour (default:red)}

\item{colour_mid}{colour scale zero colour (default:lightgrey)}

\item{colour_mid_text}{text mid colour (default:grey)}
}
\value{
Nothing
}
\description{
Plot free energy heatmap.
}
