% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_heatmap_subset.R
\name{doubledeepms__plot_heatmap_subset}
\alias{doubledeepms__plot_heatmap_subset}
\title{doubledeepms__plot_heatmap_subset}
\usage{
doubledeepms__plot_heatmap_subset(input_file, output_file, mut_subset,
  pos_subset, order_mut_subset = T, width = 10, height = 4,
  units = "in", colour_clip = 4, cluster = "both", xlab = "",
  ylab = "", xtick_labels = NULL, ytick_labels = NULL,
  colour_type = "continuous", colour_low = "blue",
  colour_high = "red", colour_mid = "white", colour_midpoint = 0,
  colour_limits = NULL, mono = F, na_colour = "grey50",
  xaxis_angle = 330, xaxis_hjust = 0, xaxis_vjust = NULL,
  xaxis_size = 5, omit_xtext = F, omit_xticks = F,
  yaxis_angle = NULL, yaxis_hjust = NULL, yaxis_vjust = NULL,
  yaxis_size = NULL, omit_ytext = F, omit_yticks = F,
  plot_title = "", input_matrix_text = NULL,
  input_matrix_point = NULL, text_size = 0.25, text_colour = "black",
  highlight_regions = NULL, x_breaks = ggplot2::waiver(),
  y_breaks = ggplot2::waiver(), show_legend = F, plot = T)
}
\arguments{
\item{input_file}{data frame with heatmap values (required)}

\item{output_file}{plot output file path}

\item{mut_subset}{vector of aa mutations to show in the heat map y axis (required)}

\item{pos_subset}{vector of domain positions to show in the heat map x axis (required)}

\item{order_mut_subset}{whether the y axis should follow the order of aa of the mut_subset (True, False, default: True)}

\item{width}{plot width in "units"}

\item{height}{plot height in "units"}

\item{units}{plot size units ("in", "cm", or "mm")}

\item{colour_clip}{maximum absolute value of colour scale}

\item{cluster}{heirarchically cluster ("none", "row", "column", "none")}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{xtick_labels}{display labels for x ticks}

\item{ytick_labels}{display labels for y ticks}

\item{colour_type}{colour scale type ("continuous", "categorical")}

\item{colour_low}{colour scale lower limit colour -- passed to scale_colour_gradient2}

\item{colour_high}{colour scale upper limit colour -- passed to scale_colour_gradient2}

\item{colour_mid}{colour scale zero colour -- passed to scale_colour_gradient2}

\item{colour_midpoint}{midpoint for colour scale -- passed to scale_colour_gradient2}

\item{colour_limits}{upper and lower value limits of colour scale -- passed to scale_colour_gradient2}

\item{mono}{use monotype font (True, False)}

\item{na_colour}{colour to use for NA values}

\item{xaxis_angle}{rotation angle for x tick labels -- passed to element_text}

\item{xaxis_hjust}{horizontal justification of x tick labels (in [0, 1]) -- passed to element_text}

\item{xaxis_vjust}{vertical justification of x tick labels (in [0, 1]) -- passed to element_text}

\item{xaxis_size}{text size of x tick labels (in pts) -- passed to element_text}

\item{omit_xtext}{omit x tick labels (True, False)}

\item{omit_xticks}{omit x ticks (True, False)}

\item{yaxis_angle}{rotation angle for y tick labels -- passed to element_text}

\item{yaxis_hjust}{horizontal justification of y tick labels (in [0, 1]) -- passed to element_text}

\item{yaxis_vjust}{vertical justification of y tick labels (in [0, 1]) -- passed to element_text}

\item{yaxis_size}{text size of y tick labels (in pts) -- passed to element_text}

\item{omit_ytext}{omit y tick labels (True, False)}

\item{omit_yticks}{omit y ticks (True, False)}

\item{plot_title}{main title for plot}

\item{input_matrix_text}{matrix of heatmap text}

\item{input_matrix_point}{matrix of heatmap points (plotted instead of tiles)}

\item{text_size}{size of heatmap text}

\item{text_colour}{colour of heatmap text}

\item{highlight_regions}{list of highlighted regions of form: list("red" = list("region1" = c(_min_, _max_), "region2" = c(_min_, _max_), ...), "blue" = list("region3" = c(_min_, _max_), ...), ...)}

\item{x_breaks}{x-axis breaks (for displaing xtick_labels)}

\item{y_breaks}{y-axis breaks (for displaing ytick_labels)}

\item{show_legend}{whether to plot legend (True, False)}

\item{plot}{whether to plot the heatmap (True, False)}
}
\value{
Nothing
}
\description{
ggplot subset and plot free energies heatmap.
}
