% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_isochore_fraction_bound.R
\name{doubledeepms__plot_isochore_fraction_bound}
\alias{doubledeepms__plot_isochore_fraction_bound}
\title{doubledeepms__plot_isochore_fraction_bound}
\usage{
doubledeepms__plot_isochore_fraction_bound(mochi_outpath, input_dt,
  RT = 0.001987 * (273 + 24), report_outpath, colour_scheme)
}
\arguments{
\item{mochi_outpath}{path to MoCHI thermo model fit results (required)}

\item{input_dt}{data.table with model free energy estimates (required)}

\item{RT}{constant (default:0.001987*(273+24))}

\item{report_outpath}{output path for scatterplots (required)}

\item{colour_scheme}{colour scheme file (default:ggplot colours)}
}
\value{
Nothing
}
\description{
Plot folding versus binding energy coloured by fraction bound showing isochores for arbitrary double mutant.
}
