% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_loadings.R
\name{doubledeepms__plot_loadings}
\alias{doubledeepms__plot_loadings}
\title{doubledeepms__plot_loadings}
\usage{
doubledeepms__plot_loadings(pca_obj, output_file, plot_categories,
  plot_colours, mono_cols = c("white", "grey", "lightgrey", "darkgrey"),
  comps = 1:5, plot_symbol = 19, points_cex = 1, width = 10,
  height = 10)
}
\arguments{
\item{pca_obj}{result of call to princomp (required)}

\item{output_file}{path to output plot (required)}

\item{plot_categories}{a character vector of category names (required)}

\item{plot_colours}{a character vector of category colours (required)}

\item{mono_cols}{a character vector of monochrome colours (to plot in background)}

\item{comps}{a numeric vector of principle component indices (default:1:5)}

\item{plot_symbol}{plot symbol (default:19)}

\item{points_cex}{points cex (default:1)}

\item{width}{plot width (default:10)}

\item{height}{plot height (default:10)}
}
\value{
Nothing
}
\description{
PCA biplot scatterplot matrix.
}
