% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_model_performance.R
\name{doubledeepms__plot_model_performance}
\alias{doubledeepms__plot_model_performance}
\title{doubledeepms__plot_model_performance}
\usage{
doubledeepms__plot_model_performance(input_dt, report_outpath)
}
\arguments{
\item{input_dt}{data.table with model free energy estimates (required)}

\item{report_outpath}{output path for scatterplots (required)}
}
\value{
Nothing
}
\description{
Plot model performance.
}
