% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_mutation_distributions.R
\name{doubledeepms__plot_mutation_distributions}
\alias{doubledeepms__plot_mutation_distributions}
\title{doubledeepms__plot_mutation_distributions}
\usage{
doubledeepms__plot_mutation_distributions(input_dt, report_outpath,
  colour_scheme)
}
\arguments{
\item{input_dt}{data.table with fitness estimates (required)}

\item{report_outpath}{output path for scatterplots (required)}

\item{colour_scheme}{colour scheme file (required)}
}
\value{
Nothing
}
\description{
Plot mutation distributions.
}
