% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/doubledeepms__plot_sturcture_examples_residues.R
\name{doubledeepms__plot_sturcture_examples_residues}
\alias{doubledeepms__plot_sturcture_examples_residues}
\title{doubledeepms__plot_sturcture_examples_residues}
\usage{
doubledeepms__plot_sturcture_examples_residues(pdb_id, pdb_view,
  domain_pos, ligand_pos, domain_backbone_color, ligand_backbone_color,
  domain_residues_color, ligand_residues_color, zoom_center = 25,
  output_file)
}
\arguments{
\item{pdb_id}{name of the pdb structure (required)}

\item{pdb_view}{vector of the pdb sturcture orientation (required)}

\item{domain_pos}{vector of domain positions to highlight in the heatmap and 3D structure (required)}

\item{ligand_pos}{vector of domain positions to highlight in the 3D structure (required)}

\item{domain_backbone_color}{color of the domain backbone in the 3D structure}

\item{ligand_backbone_color}{color of the ligand backbone in the 3D structure}

\item{domain_residues_color}{color of the domain residues in the 3D structure}

\item{ligand_residues_color}{color of the ligand residues in the 3D structure}

\item{zoom_center}{points to zoom in (negative) or out (positive) the pdb structure in points (Default: 25)}

\item{output_file}{output file path}
}
\description{
generate pymol script to get figures of pdb structures highlighting example residues
}
