% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_validation_scatter.R
\name{doubledeepms__plot_validation_scatter}
\alias{doubledeepms__plot_validation_scatter}
\title{doubledeepms__plot_validation_scatter}
\usage{
doubledeepms__plot_validation_scatter(input_dt, lit_inpath, report_outpath,
  RT = 0.001987 * (273 + 24), position_offset = 0)
}
\arguments{
\item{input_dt}{data.table with model free energy estimates (required)}

\item{lit_inpath}{path to literature free energies (required)}

\item{report_outpath}{output path for scatterplots (required)}

\item{RT}{constant (default:0.001987*(273+24))}

\item{position_offset}{residue position offset (default:0)}
}
\value{
Nothing
}
\description{
Plot correlation with validation data.
}
