% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__predict_fitness.R
\name{doubledeepms__predict_fitness}
\alias{doubledeepms__predict_fitness}
\title{doubledeepms__predict_fitness}
\usage{
doubledeepms__predict_fitness(mochi_outpath, folding_energy,
  binding_energy)
}
\arguments{
\item{mochi_outpath}{path to MoCHI thermo model fit results (required)}

\item{folding_energy}{folding energies for 3-state model (required)}

\item{binding_energy}{binding energies for 3-state model (required for binding fitness prediction)}
}
\value{
A list with folding and binding fitness predictions
}
\description{
Predict fitness from folding and binding energies for 3-state model.
}
