% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__predict_fitness_from_id.R
\name{doubledeepms__predict_fitness_from_id}
\alias{doubledeepms__predict_fitness_from_id}
\title{doubledeepms__predict_fitness_from_id}
\usage{
doubledeepms__predict_fitness_from_id(mochi_outpath, mut_id,
  RT = 0.001987 * (273 + 24), fitness_type = "folding")
}
\arguments{
\item{mochi_outpath}{path to MoCHI thermo model fit results (required)}

\item{mut_id}{mutation id (required)}

\item{RT}{constant (default:0.001987*(273+24))}

\item{fitness_type}{fitness type, either folding or binding (default:folding)}
}
\value{
fitness predictions
}
\description{
Predict fitness from id.
}
