% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__relative_angle_from_PDB.R
\name{doubledeepms__relative_angle_from_PDB}
\alias{doubledeepms__relative_angle_from_PDB}
\title{doubledeepms__relative_angle_from_PDB}
\usage{
doubledeepms__relative_angle_from_PDB(input_file, chain = "A")
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{chain}{chain id (default:A)}
}
\value{
data.table with relative angles for all residues
}
\description{
Get relative angle of c-beta to c-alpha vector w.r.t. center of mass to c-alpha vector from PDB file.
}
