% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__secondary_structure_from_PDB.R
\name{doubledeepms__secondary_structure_from_PDB}
\alias{doubledeepms__secondary_structure_from_PDB}
\title{doubledeepms__secondary_structure_from_PDB}
\usage{
doubledeepms__secondary_structure_from_PDB(input_file, chain = "A")
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{chain}{chain id (default:A)}
}
\value{
data.table with secondary structure
}
\description{
Get secondary structure.
}
