% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__temperature_factor_from_PDB.R
\name{doubledeepms__temperature_factor_from_PDB}
\alias{doubledeepms__temperature_factor_from_PDB}
\title{doubledeepms__temperature_factor_from_PDB}
\usage{
doubledeepms__temperature_factor_from_PDB(input_file, chain = "A")
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{chain}{chain id (default:A)}
}
\value{
data.table with residue-level temperature factor (B-factor) column values
}
\description{
Get residue-level data from temperature factor (B-factor) columm in PDB file.
}
