% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_allostery_plots.R
\name{doubledeepms_allostery_plots}
\alias{doubledeepms_allostery_plots}
\title{doubledeepms_allostery_plots}
\usage{
doubledeepms_allostery_plots(input_file, pdb_file_list,
  pdb_chain_query_list, annotation_list, ohm_file_list, outpath,
  colour_scheme, execute = TRUE)
}
\arguments{
\item{input_file}{path to input file (required)}

\item{pdb_file_list}{path to PDB file (required)}

\item{pdb_chain_query_list}{query chain id (required)}

\item{annotation_list}{annotations of allosteric sites (required)}

\item{ohm_file_list}{ohm output file list (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot free energy heatmaps.
}
