% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_binding_interface.R
\name{doubledeepms_binding_interface}
\alias{doubledeepms_binding_interface}
\title{doubledeepms_binding_interface}
\usage{
doubledeepms_binding_interface(input_list, outpath, colour_scheme,
  execute = TRUE)
}
\arguments{
\item{input_list}{path to MoCHI thermo model fit results (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot plots related to binding free energy and protein binding interface
}
