% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_downsampling_analysis.R
\name{doubledeepms_downsampling_analysis}
\alias{doubledeepms_downsampling_analysis}
\title{doubledeepms_downsampling_analysis}
\usage{
doubledeepms_downsampling_analysis(mochi_outpath_prefix, sample_percentage,
  outpath_suffix = "p", literature_comparison_prefix = "",
  literature_comparison_list = list(), temperature = 30, outpath,
  colour_scheme, execute = TRUE)
}
\arguments{
\item{mochi_outpath_prefix}{path prefix to MoCHI thermo model fit results (required)}

\item{sample_percentage}{sample percentages (required)}

\item{outpath_suffix}{output suffix (default:p)}

\item{literature_comparison_prefix}{literature comparison path prefix (default:empty string)}

\item{literature_comparison_list}{literature comparison file list (default:empty list)}

\item{temperature}{temperature in degrees celcuis (default:30)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Downsampling analysis.
}
