% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_eve_comparisons.R
\name{doubledeepms_eve_comparisons}
\alias{doubledeepms_eve_comparisons}
\title{doubledeepms_eve_comparisons}
\usage{
doubledeepms_eve_comparisons(eve_file_list, outpath, colour_scheme,
  execute = TRUE)
}
\arguments{
\item{eve_file_list}{EVE output file list (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot EVE comparisons.
}
