% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_fit_thermo_model.R
\name{doubledeepms_fit_thermo_model}
\alias{doubledeepms_fit_thermo_model}
\title{doubledeepms_fit_thermo_model}
\usage{
doubledeepms_fit_thermo_model(base_dir,
  mochi_output_name = "mochi__fit_tmodel_3state_sparse_dimsum128",
  tmodel_job_number = 1, tmodel_grid_search = FALSE,
  tmodel_protein = "all", tmodel_subset = "100",
  tmodel_hyperparameters, execute = TRUE)
}
\arguments{
\item{base_dir}{Base directory for all output files}

\item{mochi_output_name}{temperature in degrees celcuis (default:30)}

\item{tmodel_job_number}{Thermodynamic model fit job number: 1:final model, 2-11: monte carlo simluations for confidence intervals of model-inferred free energies (default:1)}

\item{tmodel_grid_search}{Thermodynamic model fit grid search to determine optimal hyperparameters (default:FALSE)}

\item{tmodel_protein}{Thermodynamic model fit proteins: comma-separated list of names or 'all' (default:'all')}

\item{tmodel_subset}{Thermodynamic model fit data subset: either an integer percentage of subsampled doubles (1-100) or "binding_only" or "singles_only" (default:"100")}

\item{tmodel_hyperparameters}{path table of hyperparameters}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Fit thermo models
}
