% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_fitness_heatmaps.R
\name{doubledeepms_fitness_heatmaps}
\alias{doubledeepms_fitness_heatmaps}
\title{doubledeepms_fitness_heatmaps}
\usage{
doubledeepms_fitness_heatmaps(input_file, input_file_fitness,
  input_file_MSA = NULL, domain_name, outpath, colour_scheme,
  plot_width = 10, plot_height = 4, plot_traits = c("Abundance",
  "Binding"), execute = TRUE)
}
\arguments{
\item{input_file}{path to MoCHI thermo model fit results (required)}

\item{input_file_fitness}{path to fitness data (required)}

\item{input_file_MSA}{path to MSA frequencies data (optional)}

\item{domain_name}{domain name (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{plot_width}{heatmap plot width in inches (default:10)}

\item{plot_height}{heatmap plot height in inches (default:4)}

\item{plot_traits}{traits to plot (default:Abundance, Binding)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot fitness heatmaps.
}
