% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_fitness_plots.R
\name{doubledeepms_fitness_plots}
\alias{doubledeepms_fitness_plots}
\title{doubledeepms_fitness_plots}
\usage{
doubledeepms_fitness_plots(fitness_list, val_inpath,
  structure_metrics_list, outpath, colour_scheme, execute = TRUE)
}
\arguments{
\item{fitness_list}{list of folder paths to fitness data (required)}

\item{val_inpath}{path to experimental fitness validations (required)}

\item{structure_metrics_list}{list of paths to structure metrics (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot fitness distributions and scatterplots.
}
