% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_foldx_comparisons.R
\name{doubledeepms_foldx_comparisons}
\alias{doubledeepms_foldx_comparisons}
\title{doubledeepms_foldx_comparisons}
\usage{
doubledeepms_foldx_comparisons(input_file, foldx_file_list, outpath,
  colour_scheme, execute = TRUE)
}
\arguments{
\item{input_file}{path to input file (required)}

\item{foldx_file_list}{foldx output file list (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot FoldX comparisons.
}
