% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_free_energy_scatterplots.R
\name{doubledeepms_free_energy_scatterplots}
\alias{doubledeepms_free_energy_scatterplots}
\title{doubledeepms_free_energy_scatterplots}
\usage{
doubledeepms_free_energy_scatterplots(input_list, input_MSA_list, outpath,
  colour_scheme, execute = TRUE)
}
\arguments{
\item{input_list}{path to MoCHI thermo model fit results (required)}

\item{input_MSA_list}{path to MSA frequencies data (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot free energy heatmaps.
}
