% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_interface_mechanisms.R
\name{doubledeepms_interface_mechanisms}
\alias{doubledeepms_interface_mechanisms}
\title{doubledeepms_interface_mechanisms}
\usage{
doubledeepms_interface_mechanisms(base_dir,
  input_folder_prefix = "006_doubledeepms_free_energy_heatmaps",
  domain_name, outpath, mut_subset_list, pos_subset_list, ligand_pos_list,
  pdb_id, pdb_view, zoom_pdb, plot_width = 10, plot_height = 4,
  colour_scheme, execute = TRUE)
}
\arguments{
\item{base_dir}{Base directory for all output file (required, change accordingly)}

\item{input_folder_prefix}{folder name where free energy heatmap data is (required)}

\item{domain_name}{domain name (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{mut_subset_list}{list of aa mutations to highlight in the heatmap (required)}

\item{pos_subset_list}{list of domain positions to highlight in the heatmap and 3D structure (required)}

\item{ligand_pos_list}{list of domain positions to highlight in the 3D structure (required)}

\item{pdb_id}{name of the pdb file (required)}

\item{pdb_view}{vector with pdb coordinates to orientate the 3D structure (required)}

\item{zoom_pdb}{point to zoom into the center of the pdb structure (default: 25)}

\item{plot_width}{heatmap plot width in inches (default:10)}

\item{plot_height}{heatmap plot height in inches (default:4)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot terface structures and binding free energy heatmaps subsets
}
