% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_polyphen2_comparisons.R
\name{doubledeepms_polyphen2_comparisons}
\alias{doubledeepms_polyphen2_comparisons}
\title{doubledeepms_polyphen2_comparisons}
\usage{
doubledeepms_polyphen2_comparisons(input_file, polyphen2_file,
  position_offset, uniprot_id, outpath, colour_scheme, execute = TRUE)
}
\arguments{
\item{input_file}{path to input file (required)}

\item{polyphen2_file}{polyphen2 output file (required)}

\item{position_offset}{position offset (required)}

\item{uniprot_id}{uniprot id (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot PolyPhen2 comparisons.
}
