% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_structure_metrics.R
\name{doubledeepms_structure_metrics}
\alias{doubledeepms_structure_metrics}
\title{doubledeepms_structure_metrics}
\usage{
doubledeepms_structure_metrics(input_file, outpath, pdb_file,
  pdb_RSASA_file, pdb_chain_query = "A", pdb_chain_target = "B",
  execute = TRUE)
}
\arguments{
\item{input_file}{path to MoCHI thermo model fit results (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{pdb_file}{path to PDB file (required)}

\item{pdb_RSASA_file}{path to PDB file with RSASA (required)}

\item{pdb_chain_query}{query chain id (default:A)}

\item{pdb_chain_target}{target chain id (default:B)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Add 3D structure metrics.
}
