% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms_thermo_model_results.R
\name{doubledeepms_thermo_model_results}
\alias{doubledeepms_thermo_model_results}
\title{doubledeepms_thermo_model_results}
\usage{
doubledeepms_thermo_model_results(mochi_outpath, temperature = 30,
  literature_free_energies = NA, position_offset = 0,
  folding_energy_max_sd = 1/(1.96 * 2), binding_energy_max_sd = 1/(1.96
  * 2), outpath, colour_scheme, execute = TRUE)
}
\arguments{
\item{mochi_outpath}{path to MoCHI thermo model fit results (required)}

\item{temperature}{temperature in degrees celcuis (default:30)}

\item{literature_free_energies}{path to literature free energies (default:NA)}

\item{position_offset}{residue position offset (default:0)}

\item{folding_energy_max_sd}{maximum folding energy standard deviation (default:1/(1.96*2))}

\item{binding_energy_max_sd}{maximum binding energy standard deviation (default:1/(1.96*2))}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Evaluate thermo model results.
}
