% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms.R
\name{doubledeepms}
\alias{doubledeepms}
\title{doubledeepms}
\usage{
doubledeepms(startStage = 1, stopStage = 11,
  base_dir = "/users/project/prj004631/afaure/DMS/Results/doubledeepms_proj",
  tmodel_job_number = 1, tmodel_grid_search = FALSE,
  tmodel_protein = "GRB2-SH3", tmodel_subset = 100)
}
\arguments{
\item{startStage}{Start at a specified analysis stage (default:1)}

\item{stopStage}{Stop at a specified analysis stage (default:11)}

\item{base_dir}{Base directory for all output files (default:NB private CRG server path; change accordingly)}

\item{tmodel_job_number}{Thermodynamic model fit job number: 1:final model, 2-11: monte carlo simluations for confidence intervals of model-inferred free energies (default:1)}

\item{tmodel_grid_search}{Thermodynamic model fit grid search to determine optimal hyperparameters (default:FALSE)}

\item{tmodel_protein}{Thermodynamic model fit proteins: comma-separated list of names or 'all' (default:'GRB2-SH3')}

\item{tmodel_subset}{Thermodynamic model fit data subset: either an integer percentage of subsampled doubles (1-100) or "binding_only" or "singles_only" (default:100)}
}
\value{
Nothing
}
\description{
Main analysis script.
}
