% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__persite_energy_vs_distance_plot.R
\name{doubledeepms__persite_energy_vs_distance_plot}
\alias{doubledeepms__persite_energy_vs_distance_plot}
\title{doubledeepms__persite_energy_vs_distance_plot}
\usage{
doubledeepms__persite_energy_vs_distance_plot(input_dt, outpath,
  colour_scheme, literature_sites = c(), trait_name = "folding")
}
\arguments{
\item{input_dt}{input data table (required)}

\item{outpath}{plot output path (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{literature_sites}{literature allosteric sites (default:empty vector)}

\item{trait_name}{trait name (default:folding)}
}
\value{
Allosteric positions
}
\description{
Create results folder for analysis script plots and saved objects.
}
