% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_binding_site_ROC.R
\name{doubledeepms__plot_binding_site_ROC}
\alias{doubledeepms__plot_binding_site_ROC}
\title{doubledeepms__plot_binding_site_ROC}
\usage{
doubledeepms__plot_binding_site_ROC(input_dt, outpath, colour_scheme)
}
\arguments{
\item{input_dt}{input data table (required)}

\item{outpath}{plot output path (required)}

\item{colour_scheme}{colour scheme file (required)}
}
\value{
Nothing
}
\description{
Plot binding site prediction ROC plot.
}
