% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledeepms__plot_growth_validations.R
\name{doubledeepms__plot_growth_validations}
\alias{doubledeepms__plot_growth_validations}
\title{doubledeepms__plot_fitness_replicates_cor}
\usage{
doubledeepms__plot_growth_validations(input_dt, outpath, val_inpath)
}
\arguments{
\item{input_dt}{processed data table with DiMSum fitness scores (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{val_inpath}{path to experimental fitness validations (required)}
}
\value{
Nothing
}
\description{
Create results folder for analysis script plots and saved objects.
}
